Use pre-built Adhearsion base box
This is a two-stage build. CI will build a base box, and this will be used by Vagrant for a second chef run to tweak settings for the particular environment.

Perhaps at some point in the future, the boxes can be generic enough to be used as .ova appliances without Vagrant, allowing a dependency-minimal deployment of these boxes.